drfsupport/requirements.txt
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.6
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.3
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-nested-routers==0.93.4
flake8==4.0.1
flake8-bugbear==22.4.25
flake8-commas==2.1.0
flake8-quotes==3.3.1
iniconfig==1.1.1
isort==5.10.1
kombu==5.2.4
mccabe==0.6.1
packaging==21.3
pluggy==1.0.0
prompt-toolkit==3.0.29
psycopg2==2.9.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.8
pytest==7.1.2
pytest-django==4.5.2
pytz==2021.3
redis==4.2.2
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
